feat: Add cheerleader 3D model for battle animations
Introduces a new React component for rendering a 3D cheerleader model. This model will be used to add dynamic animations to battles, enhancing the visual feedback during gameplay. The cheerleader can perform high-energy dances including jumping and arm movements, and will smoothly transition back to a neutral pose when not actively dancing.

Also corrects a minor typo in the definition of the idiom "丟三落四".